% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{AFS_SIA_Data}
\format{
A data frame with X rows and 8 columns:
\describe{
\item{d13C}{Carbon isotope composition (numeric).}
\item{d15N}{Nitrogen isotope composition (numeric).}
\item{Functional_Group}{Functional group classification (character).}
\item{source}{Source of the measurement (character).}
\item{note}{Additional notes (character).}
}
}
\source{
data from experimental design
}
\usage{
data
}
\description{
A dataset containing stable isotope analysis (SIA) data for various species.
}
\examples{
# example code
 data(Data)
}
\keyword{datasets}
