% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_resample.R
\name{diet_resample}
\alias{diet_resample}
\title{Resample within the diet distribution from simmr_output_dist.}
\usage{
diet_resample(diet_CI, num_iterations = 1000)
}
\arguments{
\item{diet_CI}{a data frame with the confidence intervals of the consumer diets produced with simmr. The output from simmr_output_dist is intended to be used for this.}

\item{num_iterations}{the number of random samples to take within the diet distribution. Default is 1000, but any positive integer can be used.}
}
\value{
a dataframe with a row for each species in the consumer diet, and num_iterations columns with a possible diet proportion for this prey.
}
\description{
The distributions are resampled in a way so that the diet proportions always sum to 1 so this output can be passed directly to Rpath.
}
\examples{
diet_resample(simmr_output_dist, num_iterations = 10000)
}
