% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diet_visualization.R
\name{diet_visualization}
\alias{diet_visualization}
\title{Quickly visualize trophic level changes from changes in diet}
\usage{
diet_visualization(
  diet_sensitivity_test_output,
  plot_type = "histogram",
  species_list = names(diet_sensitivity_test_output),
  metric = "trophic level"
)
}
\arguments{
\item{diet_sensitivity_test_output}{list of the trophic levels from the Rpath sensitivity runs. The function is designed to use the output from diet_sensitivity_test}

\item{plot_type}{type of plot to make, either "histogram", "boxplot", or "violin. Default is histogram where a seperate graph is made for each species. In boxplot, all species are plotted on the same graph.}

\item{species_list}{a vector of all species the user wants plotted. Default is all, but a subset can be specified. For larger models where only a few species are of interest, a subset is recommend.}

\item{metric}{either "trophic level" or "biomass" to specify if you want to investigate trophic level or biomass changes with the changing diets. Default is "trophic level". This should match the metric used in diet_sensitivity_test, but it onlu changes the axis name in these graphs.}
}
\value{
plot(s) showing the variability of trophic levels related to different diet matricies.
}
\description{
Using the output from diet_sensitivity_test, create histograms and boxplots of the variabliltiy in trophic levels resulting from variability in diets. This function is not intended for in-depth analysis or manuscript ready plots, but meerly for quick visualization to examine if the diet ranges significantly alter the model.
}
\examples{
diet_visualization(diet_sensitivity_test_output)
diet_visualization(diet_sensitivity_test_output, plot_type = "histogram")
diet_visualization(diet_sensitivity_test_output, plot_type = "boxplot")
diet_visualization(diet_sensitivity_test_output, species_list = c("species 1", "species 2")
}
