% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_data}
\alias{example_data}
\title{Test the sensitiity of Rpath trophic levels to changes in diets.}
\usage{
example_data()
}
\value{
a dataframe of the data used within the manuscript.
}
\description{
This function will load in the data used in the manuscript as an exampe for using the package.
}
\examples{
example_data()
}
