% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_simmr_data.R
\name{format_simmr_data}
\alias{format_simmr_data}
\title{Format stable isotope data to run through simmr}
\usage{
format_simmr_data(
  data,
  species.column,
  d13C.column,
  d15N.column,
  consumer.names,
  prey.names,
  disc.factors.c = 0.4,
  disc.factors.n = 3.4,
  disc.factors.c.sd = 1.3,
  disc.factors.n.sd = 1,
  data.type = "raw",
  d13C.column.sd = NA,
  d15N.column.sd = NA,
  group.column = NA
)
}
\arguments{
\item{data}{a dataframe with the stable isotope data. At minimum, this must include a column for each isotope (d13C and d15N), as well as a column with species names. Other columns are allowed but will not be utilized for this function.}

\item{species.column}{a character vector of the name of the column with species information}

\item{d13C.column}{a character vector of the name of the column with carbon data}

\item{d15N.column}{a character vector of the name of the column with nitrogen data}

\item{consumer.names}{a vector (or string) of the names of each consumer to create a simmr object}

\item{prey.names}{a vector (or string) of the names of each prey to include in each consumer's simmr}

\item{disc.factors.c}{a number representing the carbon discrimination factor (default is 0.4)}

\item{disc.factors.n}{a number representing the nitrogen discrimination factor (default is 3.4)}

\item{disc.factors.c.sd}{a number representing standard deviation for the carbon discrimination factor (default is 1)}

\item{disc.factors.n.sd}{a number representing standard deviation for the nitrogen discrimination factor (default is 1.3)}

\item{data.type}{a chacter of the type of data. Either "raw" if raw data is used, or "means" if the prey data are already averaged, default is raw}

\item{d13C.column.sd}{a character vector of the name of the column with carbon standard deviation data, only needed if data.type is "means", default is NA}

\item{d15N.column.sd}{a character vector of the name of the column with carbon standard deviation data, only needed if data.type is "means", default is NA}

\item{group.column}{a character vector of the name of the column with predator grouping information for simmr. The default is NA, so only specify if there is a grouping column}
}
\value{
a simmr object that can be utilized in any function within the simmr package
}
\description{
A function used to format a dataframe of raw stable isotope data for simmr runs. This removes the need to make a separate excel or .csv for each consumer and instead a consumer and their prey can be specified from a single dataframe. In this function, all prey will have the same discrimination factor, if prey have different discrimination factors please see the simmr package and format in their reccomended way. This function creates a simmr object that will be functional with the rest of simmr's functions.
}
\examples{
format_simmr_data(data, "species", "d13C", "d15N", c("species 1", "species 2"), c("species 3", "species 4", "species 5"))
}
