% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmr_diet_output_dist.R
\name{simmr_diet_output_dist}
\alias{simmr_diet_output_dist}
\title{Extract a range of diet possibilities from simmr within predefined confidence intervals}
\usage{
simmr_diet_output_dist(
  simmr.data,
  rpath.groups,
  CI.low = 0.025,
  CI.high = 0.975
)
}
\arguments{
\item{CI.low}{a number between 0 and 1 specifying the low \% range to sample from.}

\item{CI.high}{a number between 0 and 1 specifying the high \% range to sample from.}

\item{simmr_data}{the output from running a simmr model. For more infomration on this, please see the simmr documentation.}

\item{rpath_groups}{a vector of all the groups included in the rpath model. This is crucial because it will assign a percentage of 0 to every species that is in the Rpath model, but not in the simmr model so the data is formatted in a way to be passed to Rpath.}
}
\value{
a list of the diet ranges within the specified range
}
\description{
Create a list of a diet range for each species from each prey included in simmr. This can be used in a following function to create random samples within this range, or for other analysis by the user.
}
\examples{
simmr_diet_output_dist(simmr_ouput, rpath_groups, CI.low = 0.1, CI.high = 0.9)
}
